+++ examples/kvstore/handlers/verify_input.js
@@ -1,6 +1,6 @@
 /* eslint no-console:"off" */
 const { enums } = require('@arcblock/forge-proto');
-const { INSUFFICIENT_DATA } = enums.StatusCode;
+const { OK, INSUFFICIENT_DATA } = enums.StatusCode;
 
 /**
  * Each app server must implement `verifyTx` handler
@@ -12,7 +12,7 @@ const { INSUFFICIENT_DATA } = enums.StatusCode;
 module.exports = async function([tx]) {
   const kvPair = tx.itx.value;
   console.log('TxHandler.verifyTx', kvPair);
-  console.error('xxxxxxxx ERRORRROR');
+  console.error('-------- CONSOLE.ERROR ---------'); // this demonstrates all stderr are written to log files
 
   if (!kvPair.key) {
     return { code: INSUFFICIENT_DATA };
@@ -20,4 +20,6 @@ module.exports = async function([tx]) {
   if (!kvPair.value) {
     return { code: INSUFFICIENT_DATA };
   }
+
+  return { code: OK };
 };